server/pyproject.toml
[project]
name = "orpheus"
version = "0.1.0"
description = "Orpheus - AI-powered coding agent orchestration"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "fastmcp",
    "httpx",
    "morphcloud",
    "pydantic-settings",
    "sqlmodel",
    "asyncpg",
    "alembic",
    "coolname",
    "agent-client-protocol",
    "streamlit>=1.53.1",
    "psycopg2-binary>=2.9.11",
    "PyJWT[crypto]",
    "stripe>=11.0.0",
    "psycopg>=3.3.2",
    "pyyaml",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
orpheus-server = "orpheus.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orpheus"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
